Move viewpt calc back to flattened geometry.c
