+++ src/geometry.c
@@ -1,6 +1,7 @@
 #include <stdlib.h>
 #include <math.h>
 
+#include "include/coordinate.h"
 #include "include/geometry.h"
 
 float scale;
@@ -10,6 +11,12 @@ int tile_dw, tile_dh;
 int tile_nw, tile_nh;
 int cols, rows;
 int rmid, cmid;
+
+struct Coordinate screen_L = { 0 };
+struct Coordinate screen_R = { 0 };
+struct Coordinate screen_T = { 0 };
+struct Coordinate screen_B = { 0 };
+struct Coordinate viewpoint = { 0 };
 
 void geometry_initialise(float scale, float aspect, WINDOW *win)
 {
@@ -25,7 +32,26 @@ void geometry_initialise(float scale, float aspect, WINDOW *win)
     slope = ROOT3_INV * aspect;
     rmid = rows / 2;
     cmid = cols / 2;
+
+    geometry_calculate_viewpoint(coordinate_origin());
 }
+
+
+void geometry_calculate_viewpoint(struct Coordinate o)
+{
+    screen_L = coordinate_nshift(o, DIRECTION_WW, tile_nw / 2);
+    screen_R = coordinate_nshift(o, DIRECTION_EE, tile_nw / 2);
+    screen_T = coordinate_nshift(o, DIRECTION_NW, tile_nh / 4);
+    screen_B = coordinate_nshift(o, DIRECTION_SW, tile_nh / 4);
+    screen_T = coordinate_nshift(screen_T, DIRECTION_NE, tile_nh / 4);
+    screen_B = coordinate_nshift(screen_B, DIRECTION_SE, tile_nh / 4);
+
+    viewpoint = coordinate_common_ancestor(
+        coordinate_common_ancestor(screen_L, screen_R),
+        coordinate_common_ancestor(screen_T, screen_B)
+    );
+}
+
 
 float geometry_slope(void) { return slope; }
 int geometry_tile_dh(void) { return tile_dh; }
@@ -36,3 +62,4 @@ int geometry_cols(void) { return cols; }
 int geometry_rows(void) { return rows; }
 int geometry_rmid(void) { return rmid; }
 int geometry_cmid(void) { return cmid; }
+struct Coordinate geometry_viewpoint(void) { return viewpoint; }
